chore: move auth configs to config folder, remove logs, change email naming
