+++ gatsby-browser.js
@@ -1,6 +1,6 @@
 import React from 'react'
 import { Pushwoosh } from 'web-push-notifications'
-import AuthProvider from './src/components/auth/auth-context'
+import AuthProvider from './src/config/auth-context'
 import { WrapPagesWithLocaleContext } from './src/components/localization'
 import { isProduction, isLive } from './src/common/websocket/config'
 import { LocalStore } from './src/common/storage'
