+++ src/components/auth/index.ts
@@ -1,6 +1,5 @@
-import AuthProvider, { AuthContext, useAuth, useAuthValue } from './auth-context'
 import AuthLogin from './login'
 import AuthRegister from './register'
 import AuthReset from './reset'
 
-export { AuthProvider, AuthContext, useAuth, useAuthValue, AuthLogin, AuthRegister, AuthReset }
+export { AuthLogin, AuthRegister, AuthReset }
