+++ src/config/index.ts
@@ -1,0 +1,3 @@
+import AuthProvider, { AuthContext, useAuth, useAuthValue } from './auth-context'
+
+export { AuthProvider, AuthContext, useAuth, useAuthValue }
